Fix stability, add some more data
Rework how level requirements are stored/parsed.
Added prayer data
Added dragonfire protection info
Added toolType field

+++ src/main/java/binaryfissiongames/scraper/item_scraping/items/ItemWeaponInfo.java
@@ -3,9 +3,9 @@ package binaryfissiongames.scraper.item_scraping.items;
 import binaryfissiongames.scraper.scraping_steps.ScrapingData;
 
 public class ItemWeaponInfo implements ScrapingData {
-    public int damageDealt;
+    public Double damageDealt;
     public String strengthFormula;
-    public float attackSpeed;
+    public Double attackSpeed;
     public String projectile; /* arrow or bolt*/
     public boolean twoHanded;
     public String transformsInto; /*What a weapon degrades into (might only be used on kalphite whip???)*/
